Refactors unlock-publishing to use server actions
